cnpm.py/src/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
marshmallow>=3.0
pymssql>=2.2
python-dotenv>=0.21 
Flask-RESTX>=1.1.0 
flasgger
fastapi
apispec
apispec_webframeworks
flask-swagger-ui
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Authlib>=1.2.0
httpx>=0.24.0
PyJWT>=2.0.0
groq>=0.5.0